Enhance Execution Plan page with premium UI design - Add ambient background effects and gradient category tabs with animated selection - Improve TaskCard with gradient borders on hover and status indicator lines - Redesign HealthPanel with ambient glow and enhanced circular progress - Update AlertPanel with gradient backgrounds and animated icons - Enhance ExportsPanel with featured Master Plan item and staggered animations

+++ backend/app/models/task.py
@@ -1,6 +1,6 @@
 """Task database model."""
 import enum
-from sqlalchemy import String, Integer, Enum, ForeignKey, JSON
+from sqlalchemy import String, Integer, Float, Enum, ForeignKey, JSON
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.database import Base
@@ -36,7 +36,7 @@ class Task(Base):
         Enum(TaskCategory), nullable=False
     )
     priority: Mapped[int] = mapped_column(Integer, default=1, nullable=False)
-    estimated_days: Mapped[int] = mapped_column(Integer, default=1, nullable=False)
+    estimated_days: Mapped[float] = mapped_column(Float, default=1.0, nullable=False)
     status: Mapped[TaskStatus] = mapped_column(
         Enum(TaskStatus), default=TaskStatus.PENDING, nullable=False
     )
